adding basic movment
In order to get player to move forward in a boxer stance. I had to split the character movement script into two. My problem was getting the character to move forward while the character’s left shoulder being pointed forward. When I placed the character movement script on the character the movement was based on the chest pointing forward.  From my research I found that if you make an empty game object and place the character as a child of the game object. If you place the movement script on the new game object the forward direction will be based on the character new rotation. This solved the forward and backward moving directions, but then I had problems with rotation. When the character rotates, they rotate in a circle not in place. This is why I split the movement script into two. The parent hold the script that controls vertical movement (moving forward and backward), and the child object holding the rotation script which controls turning. By giving that script to the child of the game object the character now rotates in place.