made a NevMash test
Made a nav mash scene. The NavMash folder in the directory is key for contolling it. Made a new scene to test out AI script that works with the NavMash. There is code for rotation, I don't have use for but could be useful what we want the AI to detect someone only if they are in their direction.